fix bug

+++ EduQuest_Application/UseCases/Users/Queries/GetCurrentUser/GetCurrentUserQueryHandler.cs
@@ -25,7 +25,7 @@ public class GetCurrentUserQueryHandler : IRequestHandler<GetCurrentUserQuery, A
         var info = await _userRepository.GetUserByEmailAsync(request.email);
         var result = _mapper.Map<UserResponseDto>(info);
         var rank = await redisCaching.GetSortSetRankAsync("leaderboard:season1", info.Id);
-        result.statistic.Rank = (int)rank;
+        result.statistic.Rank = rank;
         return new APIResponse
         {
             IsError = false,
